+++ resources/views/back/_layouts/_editors/quill/html.blade.php
@@ -1,5 +1,5 @@
 <div data-editor data-locale="{{ $locale ?? app()->getLocale() }}" class="inset-s bg-white" id="{{ $name ?? $key }}">
-    {!! old($key, $manager->fieldValue($field, $locale ?? null)) !!}
+    {!! old($key, $field->getValue($locale ?? null)) !!}
 </div>
 
 <input name="{{ $name ?? $key }}" type="hidden" value="">