+++ resources/views/back/system/settings.blade.php
@@ -11,8 +11,13 @@
     <form action="{{ route('chief.back.settings.update') }}" id="updateForm" method="POST" role="form">
         {{ csrf_field() }}
         <input name="_method" type="hidden" value="PUT">
-        @foreach($manager->fields() as $field)
-            {!! $manager->renderField($field) !!}
+
+        @foreach($manager->editFields() as $field)
+            @formgroup
+                @slot('label',$field->getLabel())
+                @slot('description',$field->getDescription())
+                {!! $field->render() !!}
+            @endformgroup
         @endforeach
 
         <div class="stack text-right">