+++ resources/views/back/you/_form.blade.php
@@ -1,4 +1,4 @@
-@chiefformgroup(['field' => 'firstname'])
+@formgroup(['field' => 'firstname'])
     @slot('label', 'Voornaam')
     <div class="row gutter">
         <div class="column-12">
@@ -6,9 +6,9 @@
             <input id="firstName" class="input inset-s" type="text" name="firstname" value="{{ old('firstname',$user->firstname) }}">
         </div>
     </div>
-@endchiefformgroup
+@endformgroup
 
-@chiefformgroup(['field' => 'lastname'])
+@formgroup(['field' => 'lastname'])
     @slot('label', 'Achternaam')
     <div class="row gutter">
         <div class="column-12">
@@ -16,11 +16,11 @@
             <input id="lastName" class="input inset-s" type="text" name="lastname" value="{{ old('lastname',$user->lastname) }}">
         </div>
     </div>
-@endchiefformgroup
+@endformgroup
 
-@chiefformgroup(['field' => 'email'])
+@formgroup(['field' => 'email'])
     @slot('label', 'E-mail')
     @slot('description', 'Dit e-mail adres geldt tevens als jouw login.')
     <label for="email">E-mail</label>
     <input id="email" class="input inset-s" type="email" name="email" value="{{ old('email',$user->email) }}">
-@endchiefformgroup
+@endformgroup